requirement-def-service/requirements.txt
flask
fpdf
kafka-python